Chore. Move components from src to components folder. Update references.
