+++ __test__/Header.test.js
@@ -1,6 +1,6 @@
 import { shallow } from "enzyme";
 import React from "react";
-import Header from "../src/Header";
+import Header from "../components/Header";
 
 describe("#Welcome", () => {
   it("should render the text Welcome to Injection Dependent in our first text tag", () => {
